Java 8 specification version is 1.8

+++ src/test/java/uk/co/jemos/podam/test/unit/features/extensions/ExtensionsTest.java
@@ -55,7 +55,7 @@ public class ExtensionsTest extends AbstractPodamSteps {
                 podamInvocationSteps.whenIInvokeTheFactoryForClass(InputStream.class, podamFactory);
 
         String versionStr = System.getProperty("java.specification.version");
-        Integer javaSpecVersion = Integer.valueOf(versionStr);
+        Double javaSpecVersion = Double.valueOf(versionStr);
         if (javaSpecVersion < 10) {
             podamValidationSteps.theObjectShouldBeNull(inputStream2);
         } else {
